requirements.txt
Flask==1.0.2
werkzeug==0.15.1

# Application server for both development and production.
gunicorn==19.7.0

# Data and workers.
psycopg2==2.6.1
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.3.1
alembic==1.0.8

# Forms.
Flask-WTF==0.14.2
WTForms-Components==0.10.4
WTForms-Alchemy==0.16.9

#rich-text
flask_ckeditor

pytz